app/requirements.txt
Flask
gunicorn
tensorflow-cpu
numpy
pandas
scikit-learn
joblib
pretty_midi
mido
midi2audio